IRGen: @_fixed_layout classes still have resilient metadata
