+++ test/Inputs/resilient_class.swift
@@ -2,47 +2,6 @@
 import resilient_struct
 
 
-// Fixed-layout, fixed-size base class
-
-@_fixed_layout
-open class OutsideParent {
-  open var property: String = "OutsideParent.property"
-
-  open class var classProperty: String {
-    return "OutsideParent.classProperty"
-  }
-
-  public init() {
-    print("OutsideParent.init()")
-  }
-
-  open func method() {
-    print("OutsideParent.method()")
-  }
-
-  open class func classMethod() {
-    print("OutsideParent.classMethod()")
-  }
-}
-
-
-// Fixed-layout, resiliently-sized base class
-
-@_fixed_layout
-open class OutsideParentWithResilientProperty {
-  public let p: Point
-  public let s: Size
-  public let color: Int32
-
-  public final lazy var laziestNumber = 0
-
-  public init(p: Point, s: Size, color: Int32) {
-    self.p = p
-    self.s = s
-    self.color = color
-  }
-}
-
 
 // Resilient base class
 
@@ -72,21 +31,6 @@ open class ResilientOutsideParent {
 }
 
 
-// Fixed-layout, fixed-size subclass
-
-@_fixed_layout
-open class OutsideChild : OutsideParent {
-  open override func method() {
-    print("OutsideChild.method()")
-    super.method()
-  }
-
-  open override class func classMethod() {
-    print("OutsideChild.classMethod()")
-    super.classMethod()
-  }
-}
-
 
 // Resilient subclass
 
@@ -99,26 +43,6 @@ open class ResilientOutsideChild : ResilientOutsideParent {
   open override class func classMethod() {
     print("ResilientOutsideChild.classMethod()")
     super.classMethod()
-  }
-}
-
-
-// Fixed-layout, dependently-sized, generic base class
-
-@_fixed_layout
-open class GenericOutsideParent<A> {
-  open var property: A
-  public init(property: A) {
-    self.property = property
-    print("GenericOutsideParent.init()")
-  }
-
-  open func method() {
-    print("GenericOutsideParent.method()")
-  }
-
-  open class func classMethod() {
-    print("GenericOutsideParent.classMethod()")
   }
 }
 
@@ -142,27 +66,6 @@ open class ResilientGenericOutsideParent<A> {
 }
 
 
-// Fixed-layout, dependently-sized, generic subclass
-
-@_fixed_layout
-open class GenericOutsideChild<A> : GenericOutsideParent<A> {
-  public override init(property: A) {
-    print("GenericOutsideGenericChild.init(a: A)")
-    super.init(property: property)
-  }
-
-  open override func method() {
-    print("GenericOutsideChild.method()")
-    super.method()
-  }
-
-  open override class func classMethod() {
-    print("GenericOutsideChild.classMethod()")
-    super.classMethod()
-  }
-}
-
-
 // Resilient generic subclass
 
 open class ResilientGenericOutsideChild<A> : ResilientGenericOutsideParent<A> {
@@ -178,27 +81,6 @@ open class ResilientGenericOutsideChild<A> : ResilientGenericOutsideParent<A> {
 
   open override class func classMethod() {
     print("ResilientGenericOutsideChild.classMethod()")
-    super.classMethod()
-  }
-}
-
-
-// Fixed-layout, fixed-size subclass of generic class
-
-@_fixed_layout
-open class ConcreteOutsideChild : GenericOutsideParent<String> {
-  public override init(property: String) {
-    print("ConcreteOutsideChild.init(property: String)")
-    super.init(property: property)
-  }
-
-  open override func method() {
-    print("ConcreteOutsideChild.method()")
-    super.method()
-  }
-
-  open override class func classMethod() {
-    print("ConcreteOutsideChild.classMethod()")
     super.classMethod()
   }
 }
